refactor: introduce event bus and modular listeners
