+++ src/core/utils/event-bus.util.ts
@@ -1,0 +1,5 @@
+import {EventEmitter} from 'events';
+
+class EventBus extends EventEmitter {}
+
+export const eventBus = new EventBus();
