+++ src/features/audio-player/index.ts
@@ -1,0 +1,7 @@
+import {registerMessageMusicListener} from './listeners/message-music.listener';
+import {registerVoiceListener} from './listeners/voice.listener';
+
+export const initAudioPlayer = () => {
+  registerMessageMusicListener();
+  registerVoiceListener();
+};
