+++ src/features/audio-player/listeners/message-music.listener.ts
@@ -1,8 +1,9 @@
-import { Message } from 'discord.js';
-import { MusicAreas } from '@/core/mongodb/music-area.model';
+import {Message} from 'discord.js';
+import {eventBus} from '@/core/utils/event-bus.util';
+import {MusicAreas} from '@/core/mongodb/music-area.model';
 import * as Constant from '@/core/constants/index.constant';
-import { UrlService } from '@/core/services/music/url.service';
-import {logger} from "@/core/utils/logger.util";
+import {UrlService} from '@/core/services/music/url.service';
+import {logger} from '@/core/utils/logger.util';
 
 const handleYoutubeLink = async (msg: Message) => {
   const query = MusicAreas.where({ textChannelId: msg.channel.id });
@@ -48,6 +49,7 @@ const voiceCondition = async (
   }
   return true;
 };
-export default {
-  handleLink: handleYoutubeLink,
+
+export const registerMessageMusicListener = () => {
+  eventBus.on('message:created', handleYoutubeLink);
 };
