+++ src/features/moderating-message/index.ts
@@ -1,0 +1,5 @@
+import {registerMessageRestrictListener} from './listeners/restrict-message.listener';
+
+export const initModeratingMessage = () => {
+  registerMessageRestrictListener();
+};
